Fix2 绘制时修复死亡观察视角绘制摄像头所在人物自身

+++ Features/AimBot.cs
@@ -209,7 +209,7 @@ namespace Nex.Features
                     || !entity.IsAlive()
                     || entity.AddressBase == GameData.Player.AddressBase
                     || !entity.isSpottedByMask() // 无暴露 不自瞄
-                    || entity.GetDistance() == 0f // No self cam
+                    || ((int)entity.GetDistance()) == 0 // No self cam
                     )
                 {
                     continue;
